Add platformIO builds and switch to minimal spiffs image
Building with PlatformIO enables to specify libraries and their versions
in the platfromio.ini file. It makes building super easy (no need to
download the libraries manually in arduino ide)

Also switch to minimal spiffs image because this allows OTA. For now OTA
can be perfomed using the configuration portal. In the long run we could
enable OTA also via the webinterface.